HSparkRunModule

Description

Module that binds the logic of retrieving jobs from a source and executing them concurrently.

Documentation

data RunData

Constructors

RunData

Fields

env :: String

Test Dev Prod environment values = "test" "dev" or "prod"

param :: StandaloneParam

Param for interacting with specific cluster manager (Standalone/Yarn .. etc.)

sourceParam :: SqlParam

Source details (like Sql DB pool / other source resources)

runJobsReader :: ReaderT RunData IO ()

Module that binds the logic of retrieving jobs from a source and executing them concurrently.

Both Source Param with resource (sql db pool file handles .. others) and spark server param (StanadloneYarn/.. others) are injected using ReaderT

Workflow:

Extract data source and spark param

retrieve jobs using source param details

concurrently execute jobs through spark param information